Fix bonded prefix (#155)
* Fix bonded prefix

* Restore

* Fix

+++ precompile/staking/src/lib.rs
@@ -50,14 +50,14 @@ where
 	<Runtime::RuntimeCall as Dispatchable>::RuntimeOrigin: From<Option<Runtime::AccountId>>,
 	<Runtime as frame_system::Config>::AccountId: From<H160>,
 	<<Runtime as frame_system::Config>::RuntimeCall as Dispatchable>::RuntimeOrigin: OriginTrait,
-	<<Runtime as darwinia_staking::Config>::Deposit as Stake>::Item: From<u16>,
+	<<Runtime as darwinia_staking::Config>::Deposit as Stake>::Item: From<u8>,
 {
 	#[precompile::public("stake(uint256,uint256,uint8[])")]
 	fn stake(
 		handle: &mut impl PrecompileHandle,
 		ring_amount: U256,
 		kton_amount: U256,
-		deposits: Vec<u16>,
+		deposits: Vec<u8>,
 	) -> EvmResult<bool> {
 		let origin = handle.context().caller.into();
 		let deposits = deposits.into_iter().map(|i| i.into()).collect();
@@ -79,7 +79,7 @@ where
 		handle: &mut impl PrecompileHandle,
 		ring_amount: U256,
 		kton_amount: U256,
-		deposits: Vec<u16>,
+		deposits: Vec<u8>,
 	) -> EvmResult<bool> {
 		let origin = handle.context().caller.into();
 		let deposits = deposits.into_iter().map(|i| i.into()).collect();
